Le Quadri hotel
Mobile commerce is a widely used technology which provides convenience, ubiquity, and personalization. In this report, LQ-Online mobile application, and the type of m-commerce used by the mobile app, that is business to customer (B2C) model is introduced. Tools such as Flutter, Dart, Android studio, Git and Visual Studio Code are used in this report to create the innovated mobile application for LQ-Online. 
In this paper, an innovative solution for the LQ-Online mobile app has been created, where the problems of the current mobile apps have been listed out and the values created to solve the problems and create a modern solution. The benefits and challenges for both sellers and buyers are concluded, as well as design and navigation of the solution app was presented, and commercial feasibilities faced were listed out.